修改打印时打印的支付金额为应收而不是实收

+++ app/src/main/java/cn/iyunbei/handself/contract/MainContract.java
@@ -53,7 +53,7 @@ public interface MainContract {
          */
         void showEmptyView();
 
-        void setThisOrderTemp(long count);
+        void setThisOrderTemp(int count);
     }
 
     interface Presenter {
